+++ horse64/vmexec.h
@@ -52,7 +52,7 @@ typedef struct h64vmthread {
     int can_call_noasync;
 
     int kwarg_index_track_count;
-    int64_t *kwarg_index_track_map;
+    int32_t *kwarg_index_track_map;
     int arg_reorder_space_count;
     valuecontent *arg_reorder_space;
